feat: Pass payload from work of art page to component
